# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KONTi Design | Build Studio - Project Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads (sorted by score desc)
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a new public lead from the intake form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadCreateRequest"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/accept:
    post:
      operationId: acceptLead
      tags: [leads]
      summary: Accept a lead and create a Discovery-phase project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadAcceptResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameEs:
          type: string
        clientName:
          type: string
        location:
          type: string
        city:
          type: string
        phase:
          type: string
          enum: [discovery, consultation, pre_design, schematic_design, design_development, construction_documents, permits, construction, completed]
        phaseLabel:
          type: string
        phaseLabelEs:
          type: string
        phaseNumber:
          type: integer
        progressPercent:
          type: integer
        budgetAllocated:
          type: number
        budgetUsed:
          type: number
        startDate:
          type: string
        estimatedEndDate:
          type: string
        description:
          type: string
        coverImage:
          type: string
        liveCoverImage:
          type: string
          description: |
            Task #134. Derived at read time for KONTi roles only (admin/architect/team/superadmin).
            Resolves to the most recently uploaded `construction_progress` photo's URL for the
            project, falling back to `coverImage` when none exists. OMITTED for the client role.
        liveCoverUploadedAt:
          type: string
          description: |
            Task #134. ISO timestamp of the photo behind `liveCoverImage`. Present iff the live
            image was sourced from a real `construction_progress` document (i.e. NOT the
            `coverImage` fallback) — the dashboard surfaces this date in the staff alt text so
            the team can tell at a glance when the jobsite shot was taken. KONTi roles only.
        clientCoverImage:
          type: string
          description: |
            Task #134. Derived at read time for the client role only. One of five curated KONTi
            mockup images, chosen by snapping `progressPercent` to the nearest landmark
            (0 / 25 / 50 / 75 / 100) using midpoints 12.5 / 37.5 / 62.5 / 87.5. OMITTED for
            non-client roles.
        clientCoverLandmark:
          type: integer
          enum: [0, 25, 50, 75, 100]
          description: |
            Task #134. The landmark percent the `clientCoverImage` was snapped to, surfaced so
            the dashboard can render a tiny "X% milestone" pill. Client role only.
        asanaGid:
          type: string
        gammaReportUrl:
          type: string
        teamMembers:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, on_hold, completed]
        clientPhone:
          type: string
          description: Project-level contact phone for the client. Editable by the team.
        clientPostalAddress:
          type: string
          description: Project-level postal address for the client. Editable by the team.
        clientPhysicalAddress:
          type: string
          description: Project-level physical (street) address for the client. Editable by the team.
        currentStatusNote:
          type: string
          description: Plain-language "what's happening now" sentence (English) shown on the client construction card. Editable by the team. When blank the UI falls back to a deterministic phase-based summary.
        currentStatusNoteEs:
          type: string
          description: Plain-language "what's happening now" sentence (Spanish) shown on the client construction card. Editable by the team.
        squareMeters:
          type: number
          description: Project size in square meters. Edited from Project Detail and consumed (read-only) by the Contractor Calculator (B-05).
        bathrooms:
          type: integer
          description: Number of bathrooms. Project-level; consumed read-only by the Contractor Calculator (B-05).
        kitchens:
          type: integer
          description: Number of kitchens. Project-level; consumed read-only by the Contractor Calculator (B-05).
        projectType:
          type: string
          description: Project type bucket (residencial | comercial | mixto | contenedor). Project-level; consumed read-only by the Contractor Calculator (B-05).
          enum: [residencial, comercial, mixto, contenedor]
        contingencyPercent:
          type: number
          description: Default contingency percentage applied to contractor estimates for this project. Project-level; consumed read-only by the Contractor Calculator (B-05).
      required:
        - id
        - name
        - clientName
        - location
        - city
        - phase
        - phaseLabel
        - phaseLabelEs
        - phaseNumber
        - progressPercent
        - budgetAllocated
        - budgetUsed
        - startDate
        - estimatedEndDate
        - status
...
    LeadBooking:
      type: object
      properties:
        type:
          type: string
          enum: [consultation_30min, weekly_seminar]
        slot:
          type: string
        label:
          type: string
      required: [type, slot, label]
    Lead:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
          enum: [website, social, referral, media, events]
        projectType:
          type: string
          enum: [residencial, comercial, mixto, contenedor]
        location:
          type: string
        budgetRange:
          type: string
          enum: [under_150k, 150k_300k, 300k_500k, 500k_1m, over_1m]
        terrainStatus:
          type: string
          enum: [no_terrain, with_terrain, with_plans]
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        score:
          type: integer
        status:
          type: string
          enum: [new, contacted, accepted, rejected]
        booking:
          $ref: "#/components/schemas/LeadBooking"
        asanaGid:
          type: string
      required:
        - id
        - source
        - projectType
        - location
        - budgetRange
        - terrainStatus
        - contactName
        - email
        - phone
        - createdAt
        - score
        - status
    LeadCreateRequest:
      type: object
      properties:
        source:
          type: string
          enum: [website, social, referral, media, events]
        projectType:
          type: string
          enum: [residencial, comercial, mixto, contenedor]
        location:
          type: string
        budgetRange:
          type: string
          enum: [under_150k, 150k_300k, 300k_500k, 500k_1m, over_1m]
        terrainStatus:
          type: string
          enum: [no_terrain, with_terrain, with_plans]
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        booking:
          $ref: "#/components/schemas/LeadBooking"
      required:
        - source
        - projectType
        - location
        - budgetRange
        - terrainStatus
        - contactName
        - email
        - phone
    LeadAcceptResponse:
      type: object
      properties:
        lead:
          $ref: "#/components/schemas/Lead"
        project:
          $ref: "#/components/schemas/Project"
        asanaGid:
          type: string
        asanaMessage:
          type: string
        asanaMessageEs:
          type: string
      required:
        - lead
        - project
        - asanaGid
        - asanaMessage

    # ----- Task #127 Asana integration schemas -----
...